removed 'loading' from CategoryHeading plus misc fixes
- Remove 'loading' span from CategoryHeading component, now
has a placeholder percentage value that displays until data is
loaded (intention is to preserve div dimensions to stop UI jumping)

- Move numberUtil.renderPercentage -> categoryHelper.formatPercentage

- Move categoryHelper.renderCatHeadString -> CategoryHeading.formatCatHeadString

These changes needed to match existing code better (review comments)
