+++ src/util/numberUtil.ts
@@ -1,12 +1,12 @@
-// https://stackoverflow.com/a/38658925
 import { capitalizeFirstLetter } from "./stringUtil";
 
 var num =
-  "zero one two three four five six seven eight nine ten eleven twelve thirteen fourteen fifteen sixteen seventeen eighteen nineteen".split(
-    " ",
-  );
+"zero one two three four five six seven eight nine ten eleven twelve thirteen fourteen fifteen sixteen seventeen eighteen nineteen".split(
+  " ",
+);
 var tens = "twenty thirty forty fifty sixty seventy eighty ninety".split(" ");
 
+// https://stackoverflow.com/a/38658925
 function number2words(n) {
   if (n < 20) return num[n];
   var digit = n % 10;
@@ -17,8 +17,4 @@ function number2words(n) {
 
 export function numberToWords(n: number) {
   return capitalizeFirstLetter(number2words(n));
-}
-
-export const renderPercentage = (percentage: number) => {
-  return (Math.round(percentage * 10) / 10).toFixed(1)
 }
