+++ src/util/stringUtil.ts
@@ -1,4 +1,3 @@
-
 export function capitalizeFirstLetter(string) {
     return string.charAt(0).toUpperCase() + string.slice(1);
 }
